+++ android/app/test/screens/player/now_playing_modes_test.dart
@@ -1,3 +1,6 @@
+import 'dart:io';
+import 'dart:typed_data';
+
 import 'package:audio_service/audio_service.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
@@ -8,13 +11,35 @@ import 'package:mocktail/mocktail.dart';
 import 'package:heerr/api/subsonic_client.dart';
 import 'package:heerr/models/job_view.dart';
 import 'package:heerr/models/queue_response.dart';
+import 'package:heerr/offline/offline_paths.dart';
 import 'package:heerr/player/heerr_audio_handler.dart';
+import 'package:heerr/services/lyrics_service.dart';
 import 'package:heerr/player/player_provider.dart';
 import 'package:heerr/providers/queue.dart';
 import 'package:heerr/screens/player/now_playing_screen.dart';
 import 'package:heerr/utils/palette.dart';
 
 class _MockHandler extends Mock implements HeerrAudioHandler {}
+
+class _NoopAdapter implements HttpClientAdapter {
+  @override
+  Future<ResponseBody> fetch(
+    RequestOptions options,
+    Stream<Uint8List>? requestStream,
+    Future<dynamic>? cancelFuture,
+  ) async {
+    return ResponseBody.fromString(
+      '{"subsonic-response":{"status":"ok"}}',
+      200,
+      headers: <String, List<String>>{
+        'content-type': <String>['application/json'],
+      },
+    );
+  }
+
+  @override
+  void close({bool force = false}) {}
+}
 
 class _StubQueue extends Queue {
   @override
@@ -43,9 +68,23 @@ Widget _wrap(PlayerSnapshot snapshot, HeerrAudioHandler handler) {
   return ProviderScope(
     overrides: <Override>[
       audioHandlerProvider.overrideWithValue(handler),
-      subsonicDioClientProvider.overrideWith(
-        (Ref<AsyncValue<Dio>> ref) async => Dio(),
+      applicationDocumentsDirectoryProvider.overrideWith(
+        (ApplicationDocumentsDirectoryRef ref) async =>
+            Directory.systemTemp.createTempSync('heerr-modes-'),
       ),
+      subsonicDioClientProvider.overrideWith(
+        (Ref<AsyncValue<Dio>> ref) async {
+          final Dio dio = Dio(BaseOptions(baseUrl: 'http://navi.test'));
+          dio.httpClientAdapter = _NoopAdapter();
+          return dio;
+        },
+      ),
+      lyricsServiceProvider.overrideWith((LyricsServiceRef ref) async {
+        final Dio subsonic = await ref.watch(subsonicDioClientProvider.future);
+        final Dio lrcLib = Dio(BaseOptions(baseUrl: 'http://navi.test'));
+        lrcLib.httpClientAdapter = _NoopAdapter();
+        return LyricsService(subsonic, lrcLibDio: lrcLib);
+      }),
       playerSnapshotProvider.overrideWith(
         (Ref<AsyncValue<PlayerSnapshot>> ref) =>
             Stream<PlayerSnapshot>.value(snapshot),
@@ -104,6 +143,7 @@ void main() {
       (WidgetTester tester) async {
     await tester.pumpWidget(_wrap(_snap(), handler));
     await tester.pumpAndSettle();
+    await tester.ensureVisible(find.byIcon(Icons.repeat_rounded));
     await tester.tap(find.byIcon(Icons.repeat_rounded));
     await tester.pumpAndSettle();
     verify(() => handler.setRepeatMode(AudioServiceRepeatMode.all)).called(1);
@@ -115,6 +155,7 @@ void main() {
       _wrap(_snap(repeat: AudioServiceRepeatMode.all), handler),
     );
     await tester.pumpAndSettle();
+    await tester.ensureVisible(find.byIcon(Icons.repeat_rounded));
     await tester.tap(find.byIcon(Icons.repeat_rounded));
     await tester.pumpAndSettle();
     verify(() => handler.setRepeatMode(AudioServiceRepeatMode.one)).called(1);
@@ -126,6 +167,7 @@ void main() {
       _wrap(_snap(repeat: AudioServiceRepeatMode.one), handler),
     );
     await tester.pumpAndSettle();
+    await tester.ensureVisible(find.byIcon(Icons.repeat_one_rounded));
     await tester.tap(find.byIcon(Icons.repeat_one_rounded));
     await tester.pumpAndSettle();
     verify(() => handler.setRepeatMode(AudioServiceRepeatMode.none)).called(1);
@@ -135,6 +177,7 @@ void main() {
       (WidgetTester tester) async {
     await tester.pumpWidget(_wrap(_snap(), handler));
     await tester.pumpAndSettle();
+    await tester.ensureVisible(find.byIcon(Icons.shuffle_rounded));
     await tester.tap(find.byIcon(Icons.shuffle_rounded));
     await tester.pumpAndSettle();
     verify(() => handler.setShuffleMode(AudioServiceShuffleMode.all)).called(1);
@@ -146,6 +189,7 @@ void main() {
       _wrap(_snap(shuffle: AudioServiceShuffleMode.all), handler),
     );
     await tester.pumpAndSettle();
+    await tester.ensureVisible(find.byIcon(Icons.shuffle_rounded));
     await tester.tap(find.byIcon(Icons.shuffle_rounded));
     await tester.pumpAndSettle();
     verify(() => handler.setShuffleMode(AudioServiceShuffleMode.none)).called(1);